+++ src/gpu.jl
@@ -1,156 +1,1022 @@
+abstract type AbstractFaceNew <: AbstractType end
+abstract type AbstractPointNew <: AbstractType end
+abstract type AbstractTabulation <: AbstractType end
 
-function device_layout(fs::EachFace;kwargs...)
-    accessor = device_layout(fs.accessor;kwargs...)
-    EachFace(accessor)
+## ===========
+## Data-Layout
+## ===========
+
+struct FaceMinorArray{A,B} <: AbstractType
+    data::A
+    stride::B
 end
 
-function device_layout(mesh_face::MeshFace;tile_size=nothing)
-    @assert tile_size === nothing "not implemented yet"
-    mesh = GT.mesh(mesh_face)
-    D = num_dims(mesh_face)
-    face_nodes = GT.face_nodes_coalesced(mesh,D)
-    node_coordinates = GT.node_coordinates(mesh)
-    face_reference_id = GT.face_reference_id(mesh,D)
-    tabulated_values = mesh_face.space_face.workspace.values |> Tuple
-    tabulated_gradients = mesh_face.space_face.workspace.gradients |> Tuple
-    quad = mesh_face.space_face.quadrature
-    reference_weights = map(weights,GT.reference_quadratures(quad)) |> Tuple
-    Dface = 0
-    point = 0
-    MeshFaceWithDeviceLayout(
-                     Dface,
-                     point,
-                     node_coordinates,
-                     face_nodes,
-                     face_reference_id,
-                     reference_weights,
-                     tabulated_values,
-                     tabulated_gradients)
+function Adapt.adapt_structure(to,x::FaceMinorArray)
+    data = Adapt.adapt_structure(to,x.data)
+    stride = Adapt.adapt_structure(to,x.stride)
+    FaceMinorArray(data,stride)
 end
 
-function face_nodes_coalesced(mesh,d)
-    face_nodes = GT.face_nodes(mesh,d)
-    nlnodes = max_num_reference_nodes(mesh,d)
-    nfaces = num_faces(mesh,d)
-    T = Int32
-    face_nodes_coalesced = zeros(T,nfaces,nlnodes)
+Base.length(a::FaceMinorArray) = div(length(a.data),a.stride)
+
+function Base.getindex(a::FaceMinorArray,f::Integer)
+    stride = a.stride
+    offset = (f-1)*stride
+    r = (1+offset):(stride+offset)
+    view(a.data,r)
+end
+
+function face_minor_array(a::PartitionedArrays.AbstractJaggedArray)
+    # we assume that the stride is constant
+    stride = a.ptrs[2] - a.ptrs[1]
+    FaceMinorArray(a.data,stride)
+end
+
+struct FaceMajorArray{A,B} <: AbstractType
+    data::A
+    stride::B
+end
+
+Base.length(a::FaceMajorArray) = a.stride
+
+function Base.getindex(a::FaceMajorArray,f::Integer)
+    stride = a.stride
+    l = length(a.data)
+    n = div(l,stride)
+    r = f:stride:((n-1)*stride+f)
+    view(a.data,r)
+end
+
+function Adapt.adapt_structure(to,x::FaceMajorArray)
+    data = Adapt.adapt_structure(to,x.data)
+    stride = Adapt.adapt_structure(to,x.stride)
+    FaceMajorArray(data,stride)
+end
+
+function face_major_array(a::PartitionedArrays.AbstractJaggedArray)
+    # we assume that the stride is constant
+    nfaces = length(a.ptrs) - 1
+    data = similar(a.data)
+    # This is essentially a naive matrix transposition.
+    # which can be improved.
+    # This also only works on CPUs
     for face in 1:nfaces
-        nodes = face_nodes[face]
-        for lnode in 1:length(nodes)
-            node = nodes[lnode]
-            face_nodes_coalesced[face,lnode] = node
+        pini = a.ptrs[face]
+        pend = a.ptrs[face+1]-1
+        for (i,p) in enumerate(pini:pend)
+            k = (i-1)*nfaces + face
+            data[k] = a.data[p]
         end
     end
-    face_nodes_coalesced
+    FaceMajorArray(data,nfaces)
 end
 
-struct MeshFaceWithDeviceLayout{A,B,C,D,E,F} <: AbstractMeshFace
-    Dface::Int
-    point::Int
-    node_coordinates::A
-    face_nodes::B
-    face_reference_id::C
-    reference_weights::D
-    tabulated_values::E
-    tabulated_gradients::F
+## ===========
+## Mesh-related
+## ===========
+
+function each_face_new(
+    mesh::AbstractMesh, valD, quadrature::AbstractQuadrature)
+    state = tabulate(mesh,valD,quadrature)
+    each_face_new(state)
 end
 
-function at_face(x::MeshFaceWithDeviceLayout,face)
-    MeshFaceWithDeviceLayout(
-                     face,
-                     x.point,
-                     x.node_coordinates,
-                     x.face_nodes,
-                     x.face_reference_id,
-                     x.reference_weights,
-                     x.tabulated_values,
-                     x.tabulated_gradients)
+function each_face_new(quadrature::AbstractQuadrature)
+    domain = GT.domain(quadrature)
+    mesh = GT.mesh(domain)
+    num_dims = GT.num_dims(domain)
+    each_face_new(mesh,Val(num_dims),quadrature)
 end
 
-function num_faces(x::MeshFaceWithDeviceLayout)
-    length(x.face_reference_id)
+struct TabulatedMesh{A,B,C,D,E,F,G,H,I,J} <: AbstractTabulation
+    mesh::A
+    num_dims::Val{B}
+    quadrature::C
+    reference_shape_functions_value::D
+    reference_shape_functions_gradient::E
+    reference_unit_normals::F
+    face_nodes::G
+    node_coordinates::H
+    reference_weights::I
+    reference_coordinates::J
+end
+tabulated_mesh(a::TabulatedMesh) = a
+reference_shape_functions(::typeof(value),a::TabulatedMesh) = a.reference_shape_functions_value
+reference_shape_functions(::typeof(gradient),a::TabulatedMesh) = a.reference_shape_functions_gradient
+face_nodes(a::TabulatedMesh) = a.face_nodes
+node_coordinates(a::TabulatedMesh) = a.node_coordinates
+reference_weights(a::TabulatedMesh) = a.reference_weights
+quadrature(a::TabulatedMesh) = a.quadrature
+num_faces(a::TabulatedMesh) = length(face_nodes(a))  # TODO: is this correct?
+
+function Adapt.adapt_structure(to,x::TabulatedMesh)
+    mesh = nothing
+    num_dims = Adapt.adapt_structure(to,x.num_dims)
+    quadrature = nothing
+    reference_shape_functions_value = Adapt.adapt_structure(to,x.reference_shape_functions_value)
+    reference_shape_functions_gradient = Adapt.adapt_structure(to,x.reference_shape_functions_gradient)
+    reference_unit_normals = Adapt.adapt_structure(to,x.reference_unit_normals)
+    face_nodes = Adapt.adapt_structure(to,x.face_nodes)
+    node_coordinates = Adapt.adapt_structure(to,x.node_coordinates)
+    reference_weights = Adapt.adapt_structure(to,x.reference_weights)
+    reference_coordinates = Adapt.adapt_structure(to,x.reference_coordinates)
+    TabulatedMesh(
+                  mesh,
+                  num_dims,
+                  quadrature,
+                  reference_shape_functions_value,
+                  reference_shape_functions_gradient,
+                  reference_unit_normals,
+                  face_nodes,
+                  node_coordinates,
+                  reference_weights,
+                  reference_coordinates,)
 end
 
-function at_point(x::MeshFaceWithDeviceLayout,point)
-    MeshFaceWithDeviceLayout(
-                     x.Dface,
-                     point,
-                     x.node_coordinates,
-                     x.face_nodes,
-                     x.face_reference_id,
-                     x.reference_weights,
-                     x.tabulated_values,
-                     x.tabulated_gradients)
+function change_data_layout(x::TabulatedMesh;face_nodes_layout=identity)
+    TabulatedMesh(
+                  x.mesh,
+                  x.num_dims,
+                  x.quadrature,
+                  x.reference_shape_functions_value,
+                  x.reference_shape_functions_gradient,
+                  x.reference_unit_normals,
+                  face_nodes_layout(x.face_nodes),
+                  x.node_coordinates,
+                  x.reference_weights,
+                  x.reference_coordinates,)
 end
 
-function num_points(x::MeshFaceWithDeviceLayout)
-    face = x.Dface
-    rid = x.face_reference_id[face]
-    values = x.tabulated_values[rid]
-    size(values,2)
+function change_workspace_location(x::TabulatedMesh;kwargs...)
+    TabulatedMesh(
+                  x.mesh,
+                  x.num_dims,
+                  x.quadrature,
+                  x.reference_shape_functions_value,
+                  x.reference_shape_functions_gradient,
+                  x.reference_unit_normals,
+                  x.face_nodes,
+                  x.node_coordinates,
+                  x.reference_weights,
+                  x.reference_coordinates,)
 end
 
-function coordinate(f::MeshFaceWithDeviceLayout)
-    face = f.Dface
-    rid = f.face_reference_id[face]
-    point = f.point
-    values = f.tabulated_values[rid]
-    n = size(values,1)
-    Tx = eltype(f.node_coordinates)
-    Ts = eltype(values)
-    r = zero(Tx)*zero(Ts)
-    for i in 1:n
-        s = values[i,point]
-        node = f.face_nodes[face,i]
-        x = f.node_coordinates[node]
-        r += s*x
+function at_face_id(state::TabulatedMesh,face_id)
+    TabulatedFace(state,face_id,nothing,nothing)
+end
+
+function at_point_id(state::TabulatedMesh,face,i)
+    point = PointNew(face,i,nothing)
+    jacobian = GT.jacobian(point)
+    workspace = MeshPointWorkspace(jacobian)
+    PointNew(face,i,workspace)
+end
+
+struct MeshPointWorkspace{A} <: AbstractType
+    jacobian::A
+end
+jacobian(a::MeshPointWorkspace) = a.jacobian
+
+function tabulate(mesh::AbstractMesh, valD, quadrature::AbstractQuadrature)
+    D = val_parameter(valD)
+    num_dims = Val(D)
+    space = mesh_space(mesh,num_dims)
+    reference_shape_functions_value = reference_shape_functions(value,space,quadrature)
+    reference_shape_functions_gradient = reference_shape_functions(gradient,space,quadrature)
+    #TODO a single reference element for the moment
+    reference_unit_normals = first(map(normals,reference_spaces(space)))
+    face_nodes = GT.face_nodes(mesh,num_dims)
+    node_coordinates = GT.node_coordinates(mesh)
+    reference_weights = first(map(weights,reference_quadratures(quadrature)))
+    reference_coordinates = first(map(coordinates,reference_quadratures(quadrature)))
+    TabulatedMesh(
+                  mesh,
+                  num_dims,
+                  quadrature,
+                  reference_shape_functions_value,
+                  reference_shape_functions_gradient,
+                  reference_unit_normals,
+                  face_nodes,
+                  node_coordinates,
+                  reference_weights,
+                  reference_coordinates,)
+end
+
+## ===========
+## Space-related
+## ===========
+
+struct TabulatedSpace{A,B,C,D,E,F,G,H,I} <: AbstractTabulation
+    space::A
+    tabulated_mesh::B
+    reference_shape_functions_value::C
+    reference_shape_functions_gradient::D
+    reference_shape_functions_jacobian::E
+    shape_functions_value::F
+    shape_functions_gradient::G
+    shape_functions_jacobian::H
+    face_dofs::I
+end
+quadrature(a::TabulatedSpace) = quadrature(tabulated_mesh(a))
+tabulated_mesh(a::TabulatedSpace) = a.tabulated_mesh
+tabulated_space(a::TabulatedSpace) = a
+tabulated_field(a::TabulatedSpace) = a
+reference_shape_functions(::Val{:value},a::TabulatedSpace) = a.reference_shape_functions_value
+reference_shape_functions(::Val{:gradient},a::TabulatedSpace) = a.reference_shape_functions_gradient
+reference_shape_functions(::Val{:jacobian},a::TabulatedSpace) = a.reference_shape_functions_jacobian
+reference_shape_functions(f,a) = reference_shape_functions(shape_function_to_val(f), a)
+shape_functions(::Val{:value},a::TabulatedSpace) = a.shape_functions_value
+shape_functions(::Val{:gradient},a::TabulatedSpace) = a.shape_functions_gradient
+shape_functions(::Val{:jacobian},a::TabulatedSpace) = a.shape_functions_jacobian
+shape_functions(f,a) = shape_functions(shape_function_to_val(f), a)
+face_dofs(a::TabulatedSpace) = a.face_dofs
+
+function num_faces(space::TabulatedSpace)
+    if space.space !== nothing
+        num_faces(space.space)
+    elseif space.tabulated_mesh !== nothing
+        num_faces(tabulated_mesh(space))
+    else
+        num_faces(face_dofs(space))
     end
-    r
 end
 
-function ForwardDiff.jacobian(f::MeshFaceWithDeviceLayout)
-    face = f.Dface
-    rid = f.face_reference_id[face]
-    point = f.point
-    values = f.tabulated_gradients[rid]
-    n = size(values,1)
-    Tx = eltype(f.node_coordinates)
-    Ts = eltype(values)
-    r = outer(zero(Tx),zero(Ts))
-    for i in 1:n
-        s = values[i,point]
-        node = f.face_nodes[face,i]
-        x = f.node_coordinates[node]
-        r += outer(x,s)
+function tabulate(space::AbstractSpace,quadrature::AbstractQuadrature;tabulate=())
+    mesh = GT.mesh(space)
+    D = GT.num_dims(mesh)
+    tabulated_mesh = GT.tabulate(mesh,Val(D),quadrature)
+    if GT.value in tabulate
+        reference_shape_functions_value, shape_functions_value = allocate_shape_funcions(value,space,tabulated_mesh)
+    else
+        reference_shape_functions_value = nothing
+        shape_functions_value = nothing
     end
-    r
+    if ForwardDiff.gradient in tabulate
+        reference_shape_functions_gradient, shape_functions_gradient = allocate_shape_funcions(gradient,space,tabulated_mesh)
+    else
+        reference_shape_functions_gradient = nothing
+        shape_functions_gradient = nothing
+    end
+    if ForwardDiff.jacobian in tabulate
+        reference_shape_functions_jacobian, shape_functions_jacobian = allocate_shape_funcions(jacobian,space,tabulated_mesh)
+    else
+        reference_shape_functions_jacobian = nothing
+        shape_functions_jacobian = nothing
+    end
+    face_dofs = GT.face_dofs(space)
+    a = TabulatedSpace(
+                   space,
+                   tabulated_mesh,
+                   reference_shape_functions_value,
+                   reference_shape_functions_gradient,
+                   reference_shape_functions_jacobian,
+                   shape_functions_value,
+                   shape_functions_gradient,
+                   shape_functions_jacobian,
+                   face_dofs,
+                  )
 end
 
-function weight(f::MeshFaceWithDeviceLayout)
-    J = jacobian(f)
-    face = f.Dface
-    rid = f.face_reference_id[face]
-    point = f.point
-    w = f.reference_weights[rid][point]
-    dJ =change_of_measure(J) 
-    w*dJ
+function Adapt.adapt_structure(to,x::TabulatedSpace)
+    space = nothing
+    tabulated_mesh = Adapt.adapt_structure(to,x.tabulated_mesh)
+    reference_shape_functions_value = Adapt.adapt_structure(to,x.reference_shape_functions_value)
+    reference_shape_functions_gradient = Adapt.adapt_structure(to,x.reference_shape_functions_gradient)
+    reference_shape_functions_jacobian = Adapt.adapt_structure(to,x.reference_shape_functions_jacobian)
+    shape_functions_value = Adapt.adapt_structure(to,x.shape_functions_value)
+    shape_functions_gradient = Adapt.adapt_structure(to,x.shape_functions_gradient)
+    shape_functions_jacobian = Adapt.adapt_structure(to,x.shape_functions_jacobian)
+    face_dofs = Adapt.adapt_structure(to,x.face_dofs)
+    TabulatedSpace(
+                   space,
+                   tabulated_mesh,
+                   reference_shape_functions_value,
+                   reference_shape_functions_gradient,
+                   reference_shape_functions_jacobian,
+                   shape_functions_value,
+                   shape_functions_gradient,
+                   shape_functions_jacobian,
+                   face_dofs,
+                  )
 end
 
-function Adapt.adapt_structure(to,x::EachFace)
-    accessor = Adapt.adapt_structure(to,x.accessor)
-    EachFace(accessor)
+function change_data_layout(x::TabulatedSpace;face_dofs_layout=identity,kwargs...)
+    tabulated_mesh = change_data_layout(x.tabulated_mesh;kwargs...)
+    TabulatedSpace(
+                   x.space,
+                   tabulated_mesh,
+                   x.reference_shape_functions_value,
+                   x.reference_shape_functions_gradient,
+                   x.reference_shape_functions_jacobian,
+                   x.shape_functions_value,
+                   x.shape_functions_gradient,
+                   x.shape_functions_jacobian,
+                   face_dofs_layout(x.face_dofs),
+                  )
 end
 
-function Adapt.adapt_structure(to,x::MeshFaceWithDeviceLayout)
-    MeshFaceWithDeviceLayout(
-                     Adapt.adapt_structure(to,x.Dface),
-                     Adapt.adapt_structure(to,x.point),
-                     Adapt.adapt_structure(to,x.node_coordinates),
-                     Adapt.adapt_structure(to,x.face_nodes),
-                     Adapt.adapt_structure(to,x.face_reference_id),
-                     Adapt.adapt_structure(to,x.reference_weights),
-                     Adapt.adapt_structure(to,x.tabulated_values),
-                     Adapt.adapt_structure(to,x.tabulated_gradients)
-                    )
+struct GPUGlobalWorkspace{S}
+    alloc::S
 end
+
+Adapt.@adapt_structure GPUGlobalWorkspace
+
+function change_shape_functions_location(x::AbstractArray, space, workspace_location::Type{GPUGlobalWorkspace})
+    nfaces = num_faces(space)
+    ndofs = length(x)
+    GPUGlobalWorkspace(similar(x, (ndofs, nfaces)))
+end
+
+struct GPUThreadWorkspace{T, ndofs}
+end
+
+function change_shape_functions_location(x::AbstractArray, space, workspace_location::Type{GPUThreadWorkspace})
+    ndofs = length(x)
+    GPUThreadWorkspace{eltype(x), ndofs}()
+end
+
+struct GPUSharedMemWorkspace{threads_per_block, T, ndofs}
+end
+
+function change_shape_functions_location(x::AbstractArray, space, workspace_location::Type{GPUSharedMemWorkspace{threads_per_block}}) where {threads_per_block}
+    ndofs = length(x)
+    GPUSharedMemWorkspace{threads_per_block, eltype(x), ndofs}()
+end
+
+function change_shape_functions_location(x::Nothing, space, workspace_location)
+    nothing
+end
+
+function change_workspace_location(x::TabulatedSpace;workspace_location)
+    shape_functions_value = change_shape_functions_location(x.shape_functions_value, x, workspace_location)
+    shape_functions_gradient = change_shape_functions_location(x.shape_functions_gradient, x, workspace_location)
+    shape_functions_jacobian = change_shape_functions_location(x.shape_functions_jacobian, x, workspace_location)
+    TabulatedSpace(
+                   x.space,
+                   x.tabulated_mesh,
+                   x.reference_shape_functions_value,
+                   x.reference_shape_functions_gradient,
+                   x.reference_shape_functions_jacobian,
+                   shape_functions_value,
+                   shape_functions_gradient,
+                   shape_functions_jacobian,
+                   x.face_dofs,
+                  )
+end
+
+function each_face_new(space::AbstractSpace, quadrature::AbstractQuadrature;tabulate=())
+    state = GT.tabulate(space,quadrature;tabulate)
+    each_face_new(state)
+end
+
+function at_face_id(state::TabulatedSpace,i)
+    mesh_face = at_face_id(tabulated_mesh(state),i)
+    at_mesh_face(state,mesh_face)
+end
+
+function shape_function_at_mesh_face(shape_function, mesh_face)
+    shape_function
+end
+
+function shape_function_at_mesh_face(shape_function::GPUGlobalWorkspace, mesh_face)
+    view(shape_function.alloc,:,id(mesh_face))
+end
+
+import KernelAbstractions as KA
+
+function shape_function_at_mesh_face(::GPUSharedMemWorkspace{threads_per_block,T,ndofs}, mesh_face) where {threads_per_block,T,ndofs}
+    shared_alloc = KA.@localmem T (ndofs,threads_per_block)
+    thread_id = KA.@index(Local,Linear)
+    view(shared_alloc,:,thread_id)
+end
+
+function at_mesh_face(state::TabulatedSpace,mesh_face)
+    workspace = FaceWorkspace(
+        shape_function_at_mesh_face(shape_functions(Val(:value), state), mesh_face),
+        shape_function_at_mesh_face(shape_functions(Val(:gradient), state), mesh_face),
+        shape_function_at_mesh_face(shape_functions(Val(:jacobian), state), mesh_face),
+        reference_shape_functions(Val(:value), state),
+        reference_shape_functions(Val(:gradient), state),
+        reference_shape_functions(Val(:jacobian), state),
+    )
+
+    TabulatedFace(
+        state,
+        id(mesh_face),
+        workspace,
+        mesh_face,
+    )
+end
+
+function at_point_id(state::TabulatedSpace,face,i)
+    mesh_point = at_point_id(tabulated_mesh(face),face,i)
+    at_mesh_point(state,face,mesh_point)
+end
+
+struct SpacePointWorkspace{A,B,C,D,E,F,J} <: AbstractType
+    shape_functions_value::A
+    shape_functions_gradient::B
+    shape_functions_jacobian::C
+    reference_shape_functions_value::D
+    reference_shape_functions_gradient::E
+    reference_shape_functions_jacobian::F
+    jacobian::J
+end
+
+jacobian(a::SpacePointWorkspace) = a.jacobian
+reference_shape_functions(::Val{:value},a::SpacePointWorkspace) = a.reference_shape_functions_value
+reference_shape_functions(::Val{:gradient},a::SpacePointWorkspace) = a.reference_shape_functions_gradient
+reference_shape_functions(::Val{:jacobian},a::SpacePointWorkspace) = a.reference_shape_functions_jacobian
+shape_functions(::Val{:value},a::SpacePointWorkspace) = a.shape_functions_value
+shape_functions(::Val{:gradient},a::SpacePointWorkspace) = a.shape_functions_gradient
+shape_functions(::Val{:jacobian},a::SpacePointWorkspace) = a.shape_functions_jacobian
+
+function at_mesh_point(state::TabulatedSpace,face,mesh_point)
+    point = PointNew(face,id(mesh_point),mesh_point)
+    jacobian = GT.jacobian(point)
+    shape_functions_value = shape_functions(Val(:value), face)
+    shape_functions_gradient = shape_functions(Val(:gradient), face)
+    shape_functions_jacobian = shape_functions(Val(:jacobian), face)
+    reference_shape_functions_value = reference_shape_functions(Val(:value), face)
+    reference_shape_functions_gradient = reference_shape_functions(Val(:gradient), face)
+    reference_shape_functions_jacobian = reference_shape_functions(Val(:jacobian), face)
+
+    shape_functions_value = map_shape_functions(Val(:value),point,shape_functions_value)
+    shape_functions_gradient = map_shape_functions(Val(:gradient),point,shape_functions_gradient)
+    shape_functions_jacobian = map_shape_functions(Val(:jacobian),point,shape_functions_jacobian)
+
+    point_workspace = SpacePointWorkspace(
+        shape_functions_value,
+        shape_functions_gradient,
+        shape_functions_jacobian,
+        reference_shape_functions_value,
+        reference_shape_functions_gradient,
+        reference_shape_functions_jacobian,
+        jacobian,
+    )
+
+    PointNew(face,id(mesh_point),point_workspace)
+end
+
+## ===========
+## Field-related
+## ===========
+
+struct TabulatedField{A,B,C,D} <: AbstractTabulation
+    field::A
+    tabulated_space::B
+    free_values::C
+    dirichlet_values::D
+end
+quadrature(a::TabulatedField) = quadrature(tabulated_mesh(a))
+tabulated_mesh(a::TabulatedField) = tabulated_mesh(tabulated_space(a))
+tabulated_space(a::TabulatedField) = a.tabulated_space
+tabulated_field(a::TabulatedField) = a
+reference_shape_functions(f,a::TabulatedField) = reference_shape_functions(f,tabulated_space(a))
+shape_functions(f,a::TabulatedField) = shape_functions(f,tabulated_space(a))
+face_dofs(a::TabulatedField) = face_dofs(tabulated_space(a))
+free_values(a::TabulatedField) = a.free_values
+dirichlet_values(a::TabulatedField) = a.dirichlet_values
+
+function tabulate(uh::DiscreteField,quadrature::AbstractQuadrature;tabulate=())
+    space = GT.space(uh)
+    tabulated_space = GT.tabulate(space,quadrature;tabulate)
+    free_values = GT.free_values(uh)
+    dirichlet_values = GT.dirichlet_values(uh)
+    a = TabulatedField(
+                   uh,
+                   tabulated_space,
+                   free_values,
+                   dirichlet_values,
+                  )
+end
+
+function Adapt.adapt_structure(to,x::TabulatedField)
+    field = nothing
+    tabulated_space = Adapt.adapt_structure(to,x.tabulated_space)
+    free_values = Adapt.adapt_structure(to,x.free_values)
+    dirichlet_values = Adapt.adapt_structure(to,x.dirichlet_values)
+    TabulatedField(
+                   field,
+                   tabulated_space,
+                   free_values,
+                   dirichlet_values,
+                  )
+end
+
+function change_data_layout(x::TabulatedField;kwargs...)
+    tabulated_space = change_data_layout(x.tabulated_space;kwargs...)
+    TabulatedField(
+                   x.field,
+                   tabulated_space,
+                   x.free_values,
+                   x.dirichlet_values,
+                  )
+end
+
+function change_workspace_location(x::TabulatedField;kwargs...)
+    tabulated_space = change_workspace_location(x.tabulated_space;kwargs...)
+    TabulatedField(
+                   x.field,
+                   tabulated_space,
+                   x.free_values,
+                   x.dirichlet_values,
+                  )
+end
+
+function each_face_new(uh::DiscreteField, quadrature::AbstractQuadrature;tabulate=())
+    state = GT.tabulate(uh,quadrature;tabulate)
+    each_face_new(state)
+end
+
+function at_face_id(state::TabulatedField,id)
+    mesh_face = at_face_id(tabulated_mesh(state),id)
+    at_mesh_face(state,mesh_face)
+end
+
+function at_mesh_face(state::TabulatedField,mesh_face)
+    space = tabulated_space(state)
+    workspace = FaceWorkspace(
+        shape_function_at_mesh_face(shape_functions(Val(:value), space), mesh_face),
+        shape_function_at_mesh_face(shape_functions(Val(:gradient), space), mesh_face),
+        shape_function_at_mesh_face(shape_functions(Val(:jacobian), space), mesh_face),
+        reference_shape_functions(Val(:value), space),
+        reference_shape_functions(Val(:gradient), space),
+        reference_shape_functions(Val(:jacobian), space),
+    )
+    TabulatedFace(
+        state,
+        id(mesh_face),
+        workspace,
+        mesh_face,
+    )
+end
+
+function at_point_id(state::TabulatedField,face,i)
+    mesh_point = at_point_id(tabulated_mesh(tabulated_space(state)), face, i)
+    at_mesh_point(state,face,mesh_point)
+end
+
+function at_mesh_point(state::TabulatedField,face,mesh_point)
+    space_point = at_mesh_point(tabulated_space(state),workspace(face),mesh_point)
+    point = PointNew(face,id(mesh_point),space_point)
+end
+
+
+
+
+
+
+
+## ===========
+## Common functions
+## ===========
+
+struct Each{A,B,C} <: AbstractType
+    update::A
+    state::B
+    iterator::C
+end
+Base.length(iter::Each) = length(iter.iterator)
+Base.isdone(iter::Each,id) = id > length(iter)
+function Base.getindex(iter::Each,i::Integer)
+    item = iter.iterator[i]
+    iter.update(iter.state,item)
+end
+function Base.getindex(iter::Each,mesh_point::AbstractPointNew)
+    at_mesh_point(iter.state,mesh_point)
+end
+function Base.getindex(iter::Each,mesh_face::AbstractFaceNew)
+    at_mesh_face(iter.state,mesh_face)
+end
+function Base.iterate(iter::Each,i=1)
+    if Base.isdone(iter,i)
+        nothing
+    else
+        accessor = iter[i]
+        (accessor,i+1)
+    end
+end
+Base.keys(iter::Each) = LinearIndices((length(iter),))
+state(x::Each) = x.state
+iterator(x::Each) = x.iterator
+
+function Adapt.adapt_structure(to,x::Each)
+    update = Adapt.adapt_structure(to,x.update)
+    state = Adapt.adapt_structure(to,x.state)
+    iterator = Adapt.adapt_structure(to,x.iterator)
+    Each(
+         update,
+         state,
+         iterator,
+        )
+end
+
+function change_data_layout(a::Each;kwargs...)
+    Each(
+         a.update,
+         change_data_layout(a.state;kwargs...),
+         a.iterator,
+        )
+end
+
+function change_workspace_location(a::Each;kwargs...)
+    Each(
+         a.update,
+         change_workspace_location(a.state;kwargs...),
+         a.iterator,
+        )
+end
+
+struct FaceWorkspace{A,B,C,D,E,F}
+    shape_functions_value::A
+    shape_functions_gradient::B
+    shape_functions_jacobian::C
+    reference_shape_functions_value::D
+    reference_shape_functions_gradient::E
+    reference_shape_functions_jacobian::F
+end
+
+shape_functions(::Val{:value},a::FaceWorkspace) = a.shape_functions_value
+shape_functions(::Val{:gradient},a::FaceWorkspace) = a.shape_functions_gradient
+shape_functions(::Val{:jacobian},a::FaceWorkspace) = a.shape_functions_jacobian
+reference_shape_functions(::Val{:value},a::FaceWorkspace) = a.reference_shape_functions_value
+reference_shape_functions(::Val{:gradient},a::FaceWorkspace) = a.reference_shape_functions_gradient
+reference_shape_functions(::Val{:jacobian},a::FaceWorkspace) = a.reference_shape_functions_jacobian
+
+struct TabulatedFace{A,B,C,D} <: AbstractFaceNew
+    parent::A
+    id::B
+    workspace::C
+    mesh_face::D
+end
+
+parent(a::TabulatedFace) = a.parent
+id(a::TabulatedFace) = a.id
+tabulated_mesh(a::TabulatedFace) = tabulated_mesh(parent(a))
+tabulated_space(a::TabulatedFace) = tabulated_space(parent(a))
+tabulated_field(a::TabulatedFace) = tabulated_field(parent(a))
+workspace(a::TabulatedFace) = a.workspace
+at_face_id(a,id) = TabulatedFace(a,id,nothing,nothing)
+reference_shape_functions(f,a::TabulatedFace) = reference_shape_functions(f,tabulated_space(a))
+shape_functions(f,a::TabulatedFace) = shape_functions(f,workspace(a))
+
+function each_face_new(state::AbstractTabulation)
+    quadrature = GT.quadrature(state)
+    ids = faces(domain(quadrature))
+    Each(at_face_id,state,ids)
+end
+
+struct PointNew{A,B,C} <: AbstractPointNew
+    parent::A
+    id::B
+    workspace::C
+end
+parent(a::PointNew) = a.parent
+id(a::PointNew) = a.id
+workspace(a::PointNew) = a.workspace
+at_point_id(a,id) = at_point_id(parent(a),a,id)
+at_point_id(s,a,id) = PointNew(a,id)
+
+function each_point_new(face::TabulatedFace)
+    Each(at_point_id,face,1:num_points(face))
+end
+
+function at_mesh_point(state::AbstractTabulation,mesh_point::AbstractPointNew)
+    mesh_face = parent(mesh_point)
+    face = at_mesh_face(state,mesh_face)
+    at_mesh_point(state,face,mesh_point)
+end
+
+function Base.getindex(face::TabulatedFace,mesh_point::AbstractPointNew)
+    state = parent(face)
+    at_mesh_point(state,face,mesh_point)
+end
+
+function reference_shape_functions(f,space::AbstractSpace,quadrature::AbstractQuadrature)
+    rid_point_x = map(coordinates,reference_quadratures(quadrature))
+    # NB the TODOs below can be solved by introducing an extra nesting level
+    # TODO assumes same reference elems for integration and for interpolation
+    rid_to_tab = map(rid_point_x,reference_spaces(space)) do point_to_x, refface
+        collect(permutedims(tabulator(refface)(f,point_to_x))) # TODO fix this globally
+    end
+    # For the moment we assume a single element type
+    first(rid_to_tab)
+end
+
+function allocate_shape_funcions(f,space,tabulated_mesh)
+    quadrature = GT.quadrature(tabulated_mesh)
+    face = each_face_new(tabulated_mesh)[1]
+    point = each_point_new(face)[1]
+    i_ref_fun = reference_shape_functions(f,space,quadrature)
+    ref_fun = i_ref_fun[1]
+    fun = map_shape_function(f,space,1,point,ref_fun)
+    ndofs = max_num_reference_dofs(space)
+    i_ref_fun, zeros(typeof(fun),ndofs)
+end
+
+function num_points(face::AbstractFaceNew)
+    alloc = tabulated_mesh(face)
+    p_w = reference_weights(alloc)
+    length(p_w)
+end
+
+function reference_weight(point::AbstractPointNew)
+    alloc = tabulated_mesh(parent(point))
+    p_w = reference_weights(alloc)
+    p = id(point)
+    p_w[p]
+end
+
+function weight(a::AbstractPointNew)
+    w = reference_weight(a)
+    J = jacobian(a)
+    change_of_measure(J)*w
+end
+
+function coordinate(point::AbstractPointNew)
+    alloc = tabulated_mesh(parent(point))
+    i_p_ref_fun = reference_shape_functions(value,alloc)
+    node_x = node_coordinates(alloc)
+    f_i_node = face_nodes(alloc)
+    f = id(parent(point))
+    p = id(point)
+    x0 = zero(eltype(node_x))
+    s0 = zero(eltype(i_p_ref_fun))
+    o0 = outer(x0,s0)
+    J0 = zero(o0+o0)
+    sum_coordinate_no_views(node_x,f_i_node,i_p_ref_fun,p,f,J0)
+end
+
+@noinline function sum_coordinate_no_views(node_x,f_i_node::PartitionedArrays.AbstractJaggedArray,i_p_ref_fun,p,f,J0)
+    J = zero(J0)
+    i = 0
+    n = size(i_p_ref_fun,1)
+    node_ptr = f_i_node.ptrs[f]
+    while i < n
+        i += 1
+        node = f_i_node.data[node_ptr]
+        node_ptr += 1
+        x = node_x[node]
+        ref_fun = i_p_ref_fun[i,p]
+        J += x*ref_fun
+    end
+    J
+    ##TODO sum leads to much faster than hand-written loop, but why?
+    #the answer was the noinline above. But why?
+    #J = sum(i->outer(node_x[i_node[i]],i_s[i]),1:n;init=zero(J0))
+end
+
+@noinline function sum_coordinate_no_views(node_x,f_i_node::FaceMajorArray,i_p_ref_fun,p,f,J0)
+    J = zero(J0)
+    i = 0
+    n = size(i_p_ref_fun,1)
+    stride = f_i_node.stride
+    while i < n
+        ptr = i*stride + f
+        i += 1
+        node = f_i_node.data[ptr]
+        x = node_x[node]
+        ref_fun = i_p_ref_fun[i,p]
+        J += x*ref_fun
+    end
+    J
+end
+
+@noinline function sum_coordinate_no_views(node_x,f_i_node::FaceMinorArray,i_p_ref_fun,p,f,J0)
+    J = zero(J0)
+    i = 0
+    n = size(i_p_ref_fun,1)
+    stride = f_i_node.stride
+    offset = (f-one(f))*stride
+    while i < n
+        i += 1
+        ptr = i + offset
+        node = f_i_node.data[ptr]
+        x = node_x[node]
+        ref_fun = i_p_ref_fun[i,p]
+        J += x*ref_fun
+    end
+    J
+end
+
+function jacobian(point::AbstractPointNew)
+    jacobian(point,workspace(point))
+end
+
+function jacobian(point::AbstractPointNew,workspace::Nothing)
+    alloc = tabulated_mesh(parent(point))
+    i_p_ref_fun = reference_shape_functions(gradient,alloc)
+    node_x = node_coordinates(alloc)
+    f_i_node = face_nodes(alloc)
+    f = id(parent(point))
+    p = id(point)
+    x0 = zero(eltype(node_x))
+    s0 = zero(eltype(i_p_ref_fun))
+    o0 = outer(x0,s0)
+    J0 = zero(o0+o0)
+    sum_jacobian_no_views(node_x,f_i_node,i_p_ref_fun,p,f,J0)
+end
+
+function jacobian(point::AbstractPointNew,workspace)
+    jacobian(workspace)
+end
+
+@noinline function sum_jacobian_no_views(node_x,f_i_node::PartitionedArrays.AbstractJaggedArray,i_p_ref_fun,p,f,J0)
+    J = zero(J0)
+    i = 0
+    n = size(i_p_ref_fun,1)
+    node_ptr = f_i_node.ptrs[f]
+    while i < n
+        i += 1
+        node = f_i_node.data[node_ptr]
+        node_ptr += 1
+        x = node_x[node]
+        ref_fun = i_p_ref_fun[i,p]
+        J += outer(x,ref_fun)
+    end
+    J
+    ##TODO sum leads to much faster than hand-written loop, but why?
+    #the answer was the noinline above. But why?
+    #J = sum(i->outer(node_x[i_node[i]],i_s[i]),1:n;init=zero(J0))
+end
+
+@noinline function sum_jacobian_no_views(node_x,f_i_node::FaceMajorArray,i_p_ref_fun,p,f,J0)
+    J = zero(J0)
+    i = 0
+    n = size(i_p_ref_fun,1)
+    stride = f_i_node.stride
+    while i < n
+        ptr = i*stride + f
+        i += 1
+        node = f_i_node.data[ptr]
+        x = node_x[node]
+        ref_fun = i_p_ref_fun[i,p]
+        J += outer(x,ref_fun)
+    end
+    J
+end
+
+@noinline function sum_jacobian_no_views(node_x,f_i_node::FaceMinorArray,i_p_ref_fun,p,f,J0)
+    J = zero(J0)
+    i = 0
+    n = size(i_p_ref_fun,1)
+    stride = f_i_node.stride
+    offset = (f-one(f))*stride
+    while i < n
+        i += 1
+        ptr = i + offset
+        node = f_i_node.data[ptr]
+        x = node_x[node]
+        ref_fun = i_p_ref_fun[i,p]
+        J += outer(x,ref_fun)
+    end
+    J
+end
+
+function shape_functions(f,point::AbstractPointNew)
+    shape_functions(f,workspace(point))
+end
+
+function reference_shape_functions(f,point::AbstractPointNew)
+    reference_shape_functions(f,workspace(point))
+end
+
+function map_shape_functions(f,point::AbstractPointNew,alloc)
+    space = workspace(point)
+    i_p_ref_fun = reference_shape_functions(f,parent(point))
+    p = id(point)
+    ndofs = size(i_p_ref_fun,1)
+    i = 0
+    while i < ndofs
+        i += 1
+        ref_fun = i_p_ref_fun[i,p]
+        fun = map_shape_function(f,space,i,point,ref_fun)
+        alloc[i] = fun
+    end
+    alloc
+end
+
+function map_shape_functions(f, point::AbstractPointNew, ::GPUThreadWorkspace{T,max_ndofs}) where {T,max_ndofs}
+    space = workspace(point)
+    i_p_ref_fun = reference_shape_functions(f,parent(point))
+    p = id(point)
+    ndofs = size(i_p_ref_fun,1)
+    map(enumerate_static(zero(SVector{max_ndofs,T}))) do (i, v)
+        if i <= ndofs
+            map_shape_function(f, space, i, point, i_p_ref_fun[i, p])
+        else
+            zero(v)
+        end        
+    end
+end
+
+function map_shape_functions(f,point::AbstractPointNew,::Nothing)
+    nothing
+end
+
+function dofs(face::AbstractFaceNew)
+    alloc = GT.tabulated_space(face)
+    f_i_dof = face_dofs(alloc)
+    f_i_dof[id(face)]
+end
+
+function num_dofs(face::AbstractFaceNew)
+    alloc = GT.tabulated_space(face)
+    f_i_dof = face_dofs(alloc)
+    f = id(face)
+    num_dofs_impl(f_i_dof,f)
+end
+
+function num_dofs_impl(f_i_dof::PartitionedArrays.AbstractJaggedArray,f)
+    f_i_dof.ptrs[f+1] - f_i_dof.ptrs[f]
+end
+
+function num_dofs_impl(f_i_dof::FaceMajorArray,f)
+    div(length(f_i_dof.data),f_i_dof.stride)
+end
+
+function num_dofs_impl(f_i_dof::FaceMinorArray,f)
+    f_i_dof.stride
+end
+
+function field(f,point::AbstractPointNew)
+    alloc = GT.tabulated_field(parent(point))
+    i_fun = shape_functions(f,point)
+    @assert i_fun !== nothing
+    free_dof_val = free_values(alloc)
+    diri_dof_val = dirichlet_values(alloc)
+    f_i_dof = face_dofs(alloc)
+    f = id(parent(point))
+    x0 = zero(eltype(free_dof_val))
+    s0 = zero(eltype(i_fun))
+    o0 = x0*s0
+    J0 = zero(o0+o0)
+    sum_field_no_views(free_dof_val,diri_dof_val,f_i_dof,i_fun,f,J0)
+end
+
+@noinline function sum_field_no_views(free_dof_val,diri_dof_val,f_i_dof::PartitionedArrays.AbstractJaggedArray,i_fun,f,J0)
+    J = zero(J0)
+    i = 0
+    n = length(i_fun)
+    dof_ptr = f_i_dof.ptrs[f]
+    while i < n
+        i += 1
+        dof = f_i_dof.data[dof_ptr]
+        dof_ptr += 1
+        if dof < 0
+            val = diri_dof_val[-dof]
+        else
+            val = free_dof_val[dof]
+        end
+        fun = i_fun[i]
+        J += val*fun
+    end
+    J
+end
+
+@noinline function sum_field_no_views(free_dof_val,diri_dof_val,f_i_dof::FaceMajorArray,i_fun,f,J0)
+    J = zero(J0)
+    i = 0
+    n = length(i_fun)
+    stride = f_i_dof.stride
+    while i < n
+        ptr = i*stride + f
+        i += 1
+        dof = f_i_dof.data[ptr]
+        if dof < 0
+            val = diri_dof_val[-dof]
+        else
+            val = free_dof_val[dof]
+        end
+        fun = i_fun[i]
+        J += val*fun
+    end
+    J
+end
+
+@noinline function sum_field_no_views(free_dof_val,diri_dof_val,f_i_dof::FaceMinorArray,i_fun,f,J0)
+    J = zero(J0)
+    i = 0
+    n = length(i_fun)
+    stride = f_i_dof.stride
+    offset = (f-one(f))*stride
+    while i < n
+        i += 1
+        ptr = i + offset
+        dof = f_i_dof.data[ptr]
+        if dof < 0
+            val = diri_dof_val[-dof]
+        else
+            val = free_dof_val[dof]
+        end
+        fun = i_fun[i]
+        J += val*fun
+    end
+    J
+end
+
+
+
+
